[NCL-5109] Change WebSocket URL
New websocket url is: `ws://<url>/pnc-rest-new/notifications`
