+++ notifications/src/main/java/org/jboss/pnc/notification/MessageType.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rest.notifications.websockets;
+package org.jboss.pnc.notification;
 
 /**
  * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
