+++ notifications/src/test/java/org/jboss/pnc/notification/SessionBasedAttachedClientTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rest.notifications.websockets;
+package org.jboss.pnc.notification;
 
 import org.jboss.pnc.spi.notifications.Notifier;
 import org.junit.Test;
